+++ src/Order/Models/Billpayer.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 
 namespace Vanilo\Order\Models;
 
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Model;
 use Konekt\Address\Models\AddressProxy;
 use Vanilo\Contracts\Address;
@@ -25,9 +26,19 @@ use Vanilo\Order\Contracts\Billpayer as BillpayerContract;
  * work temporarily as of v0.1. Probably will be moved to the
  * billing module or another module, if it survives at all
  */
+
+/**
+ * @property-read string $country_id
+ * @property-read string|null $province_id
+ * @property-read string|null $postalcode
+ * @property-read string $city
+ * @property-read string $street_address
+ * @property-read string|null $address2
+ * @property-read string|null $access_code
+ */
 class Billpayer extends Model implements BillpayerContract, VaniloBillpayerContract
 {
-    protected $guarded = ['id', 'address_id'];
+    protected $guarded = ['id', 'created_at', 'updated_at'];
 
     public function isEuRegistered(): bool
     {
@@ -96,5 +107,41 @@ class Billpayer extends Model implements BillpayerContract, VaniloBillpayerContr
     public function getFullName(): string
     {
         return $this->getFirstName() . ' ' . $this->getLastName();
+    }
+
+    protected function countryId(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->getCountryCode());
+    }
+
+    protected function provinceId(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->province_id);
+    }
+
+    protected function postalcode(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->getPostalCode());
+    }
+
+    protected function city(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->getCity());
+    }
+
+    /** @todo this would collide with the address() thus the `address` -> `street_address` workaround */
+    protected function streetAddress(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->getAddress());
+    }
+
+    protected function address2(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->getAddress2());
+    }
+
+    protected function accessCode(): Attribute
+    {
+        return Attribute::get(fn ($value) => $this->getBillingAddress()->access_code);
     }
 }