+++ src/Order/Tests/BillpayerTest.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Vanilo\Order\Tests;
 
 use Vanilo\Order\Models\Billpayer;
+use Vanilo\Order\Tests\Dummies\Address;
 
 class BillpayerTest extends TestCase
 {
@@ -26,5 +27,32 @@ class BillpayerTest extends TestCase
         $this->assertIsBool($billPayer->isOrganization());
         $this->assertIsBool($billPayer->isEuRegistered());
         $this->assertIsBool($billPayer->isIndividual());
+    }
+
+    /** @test */
+    public function it_proxies_its_missing_address_fields_to_its_underlying_address_object()
+    {
+        $address = Address::create([
+            'name' => 'Sample Name',
+            'country_id' => 'NL',
+            'province_id' => '1',
+            'postalcode' => '1234 AB',
+            'city' => 'Buiswijk',
+            'address' => 'Line 1',
+            'address2' => 'Line 2',
+            'access_code' => 'CODE CODE',
+        ])->fresh();
+
+        $billPayer = Billpayer::create([
+            'address_id' => $address->id,
+        ]);
+
+        $this->assertEquals('NL', $billPayer->country_id);
+        $this->assertEquals('1', $billPayer->province_id);
+        $this->assertEquals('1234 AB', $billPayer->postalcode);
+        $this->assertEquals('Buiswijk', $billPayer->city);
+        $this->assertEquals('Line 1', $billPayer->street_address);
+        $this->assertEquals('Line 2', $billPayer->address2);
+        $this->assertEquals('CODE CODE', $billPayer->access_code);
     }
 }