+++ Changelog.md
@@ -6,6 +6,7 @@
 ##### 2024-XX-YY
 
 - Added the `priceGreaterThan`, `priceLessThan` and `priceBetween` methods to the ProductSearch class by [Matima](https://github.com/mahdirezaei-dev) in [#176](https://github.com/vanilophp/framework/pull/176)
+- Added the `Macroable` trait to the `ProductSearch` class
 - Changed the offline payment gateway's icon from a circle to a plug+x
 
 ## 4.0.1