+++ Changelog.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ##### 2024-XX-YY
 
+- Added the `includeVariants()` method to the ProductSearch class
 - Added the `priceGreaterThan`, `priceLessThan` and `priceBetween` methods to the ProductSearch class by [Matima](https://github.com/mahdirezaei-dev) in [#176](https://github.com/vanilophp/framework/pull/176)
 - Added the `Macroable` trait to the `ProductSearch` class
 - Added the unidirectional links feature